refactor: rename addressInfo -> account. Redirecting to a new page if the account is uninitialized
