--- examples/wallet/app/containers/HomePage.js
@@ -1,13 +1,0 @@
-// @flow
-import React, { Component } from 'react';
-import Home from '../components/Home';
-
-type Props = {};
-
-export default class HomePage extends Component<Props> {
-  props: Props;
-
-  render() {
-    return <Home />;
-  }
-}
